Breaking change in multipartBoundary, fixed header manipulation
Previously the multipartBoundary pointer was used to change the original header to include the two required hyphens but that manipulated the original header buffer. This manipulation was removed but the pointer moved by a wrong amount. With this commit multipartBoundary will always point to the start of the boundary value but users need to prepend the two hyphens manually.